'Update links and routes for Home and Cadastro Aluno components'
This diff includes changes to the Navbar component, the Routes component, and the CadastroAluno component. In the Navbar component, the href attribute of the Home link has been changed from "/" to "home", and the href attribute of the Cadastro Aluno link has been changed from "#" to "cadastro-aluno". In the Routes component, the Route components for Home and CadastroAluno have been updated to use the element prop instead of the component prop, and the exact prop has been added to both Route components. In the CadastroAluno component, a comment has been added to the Card Body div.