"Update routing and navbar to use Routes and remove hash in paths"
This diff updates the routing configuration to use the `Routes` component instead of `Switch` and removes the hash (#) from the paths. It also updates the `Navbar` component to use `HashRouter` and adjusts the links to match the new path format. Additionally, it updates the `rotas.js` file to import `Route` and `Routes` from `react-router-dom` instead of `HashRouter`. Finally, it removes the dropdown menu from the `Navbar` component.